+++ CHANGELOG.md
@@ -14,6 +14,8 @@ At release time, move the entries from here to a new release section.
 
 ### Additions
 
+- Parameterised the `targetRevision` for the ArgoCD application using a Terraform variable.
+
 ### Fixes
 
 ### Changes